scout-backend/requirements.txt
# SCOUT Backend Dependencies - Step 1: Basic FastAPI Setup
# Core Framework (minimal for testing)
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Strands Agents SDK for agent development
strands-agents==1.7.1
strands-agents-tools[mem0_memory,a2a]==1.7.1

# AWS dependencies
boto3==1.34.0
botocore==1.34.0

# Additional utilities
pydantic==2.5.0
python-dotenv==1.0.0
requests>=2.31.0
websocket-client>=1.6.0

# PDF Conversion
markdown-pdf==1.2.0
PyMuPDF==1.23.26
